OwnerShip and auth routes added

+++ app/Http/Controllers/ListingController.php
@@ -49,6 +49,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
 
@@ -59,12 +61,21 @@ class ListingController extends Controller
     //show edit form
     public function edit(Listing $listing)
     {
+        //make sure logged in user is ownner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         return view('listings.edit', ['listing' => $listing]);
     }
 
     //Upadte Listing Data
     public function update(Request $request, Listing $listing)
     {
+        //make sure logged in user is ownner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
 
         $formFields = $request->validate([
             'title' => 'required',
@@ -91,7 +102,18 @@ class ListingController extends Controller
     //Delete Listing
     public function destroy(Listing $listing)
     {
+        //make sure logged in user is ownner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+    //manage Listing
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
